Split SEV-SNP attestation crates
