+++ dstack-attest/Cargo.toml
@@ -11,23 +11,21 @@ license.workspace = true
 
 [dependencies]
 anyhow.workspace = true
-base64.workspace = true
 cc-eventlog.workspace = true
 rmp-serde.workspace = true
-reqwest = { workspace = true, features = ["blocking"] }
 dcap-qvl.workspace = true
 dstack-types.workspace = true
 ez-hash.workspace = true
 fs-err.workspace = true
 hex.workspace = true
 hex_fmt.workspace = true
-libc.workspace = true
 or-panic.workspace = true
 scale = { workspace = true, features = ["derive"] }
+sev-snp-attest.workspace = true
+sev-snp-qvl.workspace = true
 serde.workspace = true
 serde-human-bytes.workspace = true
 serde_json.workspace = true
-sev.workspace = true
 sha2.workspace = true
 sha3.workspace = true
 tdx-attest.workspace = true
